Tested single player in multiplayer mode.
