+++ hw07/src/main/kotlin/ru/hse/spb/client/RoguelikeClient.kt
@@ -1,24 +1,9 @@
-/*
- * Copyright 2015 The gRPC Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package ru.hse.spb.client
 
 import io.grpc.ManagedChannel
 import io.grpc.ManagedChannelBuilder
 import io.grpc.stub.StreamObserver
+import ru.hse.spb.RoguelikeSinglePlayerApplication
 import ru.hse.spb.controller.Controller
 import ru.hse.spb.model.Map
 import ru.hse.spb.model.Model
@@ -113,8 +98,12 @@ internal constructor(private val channel: ManagedChannel) {
         @Throws(Exception::class)
         @JvmStatic
         fun main(args: Array<String>) {
-            if (args.size != 2) {
+            if (args.size > 2) {
                 printUsage()
+                return
+            }
+            if (args.size < 2) {
+                RoguelikeSinglePlayerApplication.start()
                 return
             }
             val client = RoguelikeClient(args[0], args[1].toInt())
@@ -142,7 +131,7 @@ internal constructor(private val channel: ManagedChannel) {
         }
 
         private fun printUsage() {
-            println("Args: <ip> <port>")
+            println("Args: <ip> <port> or empty. Last case is single player mode")
         }
     }
 }
